# apps/api/src/orders/orders.controller.ts
import {
  Body,
  Controller,
  Get,
  Headers,
  HttpCode,
  Param,
  Post,
  Query,
  UseGuards,
} from '@nestjs/common';
import {
  ApiBearerAuth,
  ApiCreatedResponse,
  ApiHeader,
  ApiOkResponse,
  ApiQuery,
  ApiTags,
} from '@nestjs/swagger';
import { AdmissionGuard } from '../admission/admission.guard';
import { CurrentUser, type RequestUser } from '../auth/current-user.decorator';
import { JwtAuthGuard, OptionalJwtAuthGuard } from '../auth/guards';
import { CreateOrderDto } from './dto/create-order.dto';
import { OrdersService } from './orders.service';

import { MyTicketDto, OrderDetailDto } from './dto/order-response.dto';

@ApiTags('orders')
@Controller()
export class OrdersController {
  constructor(private readonly orders: OrdersService) {}

  @Post('events/:eventId/orders')
  @ApiCreatedResponse({ type: OrderDetailDto })
  @HttpCode(201)
  @UseGuards(OptionalJwtAuthGuard, AdmissionGuard)
  @ApiHeader({ name: 'idempotency-key', required: false })
  @ApiHeader({ name: 'x-hold-key', required: false })
  @ApiHeader({ name: 'x-admission-token', required: false })
  async create(
    @Param('eventId') eventId: string,
    @Body() dto: CreateOrderDto,
    @CurrentUser() user: RequestUser | null,
    @Headers('idempotency-key') idempotencyKey?: string,
    @Headers('x-hold-key') holderKey?: string,
  ) {
    const { order } = await this.orders.create({
      eventId,
      dto,
      buyerUserId: user?.id ?? null,
      idempotencyKey: idempotencyKey?.slice(0, 120) ?? null,
      holderKey: holderKey ?? null,
    });
    return order;
  }

  @Get('orders/:id')
  @ApiOkResponse({ type: OrderDetailDto })
  @UseGuards(OptionalJwtAuthGuard)
  @ApiQuery({ name: 'token', required: false })
  getById(
    @Param('id') id: string,
    @CurrentUser() user: RequestUser | null,
    @Query('token') token?: string,
  ) {
    return this.orders.getById(id, {
      userId: user?.id ?? null,
      guestToken: token ?? null,
    });
  }

  @Get('me/tickets')
  @ApiOkResponse({ type: [MyTicketDto] })
  @UseGuards(JwtAuthGuard)
  @ApiBearerAuth()
  listMyTickets(@CurrentUser() user: RequestUser) {
    return this.orders.listMyTickets(user.id);
  }
}