# apps/api/src/seatmaps/seatmaps.controller.ts
import {
  Body,
  Controller,
  Get,
  Headers,
  Param,
  Post,
  UseGuards,
} from '@nestjs/common';
import {
  ApiBearerAuth,
  ApiCreatedResponse,
  ApiHeader,
  ApiOkResponse,
  ApiProperty,
  ApiPropertyOptional,
  ApiTags,
} from '@nestjs/swagger';
import { Type } from 'class-transformer';
import {
  ArrayMaxSize,
  ArrayMinSize,
  IsArray,
  IsInt,
  IsOptional,
  IsString,
  Length,
  Max,
  Min,
  ValidateNested,
} from 'class-validator';
import { AdmissionGuard } from '../admission/admission.guard';
import { CurrentUser, type RequestUser } from '../auth/current-user.decorator';
import { JwtAuthGuard } from '../auth/guards';
import { SeatmapsService } from './seatmaps.service';

export class SeatMapSectionDto {
  @ApiProperty({ example: 'Front' })
  @IsString()
  @Length(1, 20)
  name!: string;

  @ApiProperty({ minimum: 1, maximum: 26 })
  @IsInt()
  @Min(1)
  @Max(26)
  rows!: number;

  @ApiProperty({ minimum: 1, maximum: 30 })
  @IsInt()
  @Min(1)
  @Max(30)
  cols!: number;

  @ApiProperty({ example: 'Front zone' })
  @IsString()
  @Length(1, 80)
  tierName!: string;

  @ApiProperty({
    minimum: 0,
    maximum: 100_000_000,
    description: 'Price in satang; 0 = free',
  })
  @IsInt()
  @Min(0)
  @Max(100_000_000)
  priceSatang!: number;

  @ApiPropertyOptional({
    minimum: 0,
    maximum: 80,
    description: 'Grid column of the section origin (editor layout)',
  })
  @IsOptional()
  @IsInt()
  @Min(0)
  @Max(80)
  x?: number;

  @ApiPropertyOptional({
    minimum: 0,
    maximum: 200,
    description: 'Grid row of the section origin (editor layout)',
  })
  @IsOptional()
  @IsInt()
  @Min(0)
  @Max(200)
  y?: number;
}

export class CreateSeatMapDto {
  @ApiProperty({ type: [SeatMapSectionDto] })
  @IsArray()
  @ArrayMinSize(1)
  @ArrayMaxSize(12)
  @ValidateNested({ each: true })
  @Type(() => SeatMapSectionDto)
  sections!: SeatMapSectionDto[];
}

import { SeatMapDataDto } from './dto/seat-map-response.dto';

@ApiTags('seat-maps')
@Controller('events/:eventId/seat-map')
export class SeatmapsController {
  constructor(private readonly seatmaps: SeatmapsService) {}

  @Post()
  @ApiCreatedResponse({ type: SeatMapDataDto })
  @UseGuards(JwtAuthGuard)
  @ApiBearerAuth()
  create(
    @CurrentUser() user: RequestUser,
    @Param('eventId') eventId: string,
    @Body() dto: CreateSeatMapDto,
  ) {
    return this.seatmaps.create(eventId, user.id, dto);
  }

  @Get()
  @ApiOkResponse({ type: SeatMapDataDto })
  @UseGuards(AdmissionGuard)
  @ApiHeader({ name: 'x-hold-key', required: false })
  @ApiHeader({ name: 'x-admission-token', required: false })
  get(
    @Param('eventId') eventId: string,
    @Headers('x-hold-key') holderKey?: string,
  ) {
    return this.seatmaps.getForEvent(eventId, holderKey ?? null);
  }
}