# apps/api/src/team/dto/team-member.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsEmail, IsIn } from 'class-validator';

export class AddTeamMemberDto {
  @ApiProperty()
  @IsEmail()
  email!: string;

  @ApiProperty({ enum: ['manager', 'staff'] })
  @IsIn(['manager', 'staff'])
  role!: 'manager' | 'staff';
}

export class UpdateTeamMemberDto {
  @ApiProperty({ enum: ['manager', 'staff'] })
  @IsIn(['manager', 'staff'])
  role!: 'manager' | 'staff';
}

export class TeamMemberDto {
  id: string;
  email: string;
  role: string;
  linked: boolean;
  displayName: string | null;
  createdAt: string;
}