# packages/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/events/{eventId}/team": {
      "get": {
        "operationId": "TeamController_list",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamMemberDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "team"
        ]
      },
      "post": {
        "operationId": "TeamController_add",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddTeamMemberDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamMemberDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "team"
        ]
      }
    },
    "/api/events/{eventId}/team/{memberId}": {
      "patch": {
        "operationId": "TeamController_changeRole",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTeamMemberDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamMemberDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "team"
        ]
      },
      "delete": {
        "operationId": "TeamController_remove",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "team"
        ]
      }
    },
...
    "/api/events/{eventId}/seat-map": {
      "post": {
        "operationId": "SeatmapsController_create",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSeatMapDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeatMapDataDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "seat-maps"
        ]
      },
      "get": {
        "operationId": "SeatmapsController_get",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-hold-key",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-admission-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeatMapDataDto"
                }
              }
            }
          }
        },
        "tags": [
          "seat-maps"
        ]
      }
    },
    "/api/events/{eventId}/orders": {
      "post": {
        "operationId": "OrdersController_create",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-hold-key",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-admission-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetailDto"
                }
              }
            }
          }
        },
        "tags": [
          "orders"
        ]
      }
    },
    "/api/orders/{id}": {
      "get": {
        "operationId": "OrdersController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetailDto"
                }
              }
            }
          }
        },
        "tags": [
          "orders"
        ]
      }
    },
    "/api/me/tickets": {
      "get": {
        "operationId": "OrdersController_listMyTickets",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MyTicketDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "OpenSeat API",
    "description": "REST API for the OpenSeat ticketing platform",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "TeamMemberDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "linked": {
            "type": "boolean"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "linked",
          "displayName",
          "createdAt"
        ]
      },
      "AddTeamMemberDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "enum": [
              "manager",
              "staff"
            ],
            "type": "string"
          }
        },
        "required": [
          "email",
          "role"
        ]
      },
      "UpdateTeamMemberDto": {
        "type": "object",
        "properties": {
          "role": {
            "enum": [
              "manager",
              "staff"
            ],
            "type": "string"
          }
        },
        "required": [
          "role"
        ]
      },
...
      "SeatMapSectionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Front"
          },
          "rows": {
            "type": "number",
            "minimum": 1,
            "maximum": 26
          },
          "cols": {
            "type": "number",
            "minimum": 1,
            "maximum": 30
          },
          "tierName": {
            "type": "string",
            "example": "Front zone"
          },
          "priceSatang": {
            "type": "number",
            "minimum": 0,
            "maximum": 100000000,
            "description": "Price in satang; 0 = free"
          },
          "x": {
            "type": "number",
            "minimum": 0,
            "maximum": 80,
            "description": "Grid column of the section origin (editor layout)"
          },
          "y": {
            "type": "number",
            "minimum": 0,
            "maximum": 200,
            "description": "Grid row of the section origin (editor layout)"
          }
        },
        "required": [
          "name",
          "rows",
          "cols",
          "tierName",
          "priceSatang"
        ]
      },
      "CreateSeatMapDto": {
        "type": "object",
        "properties": {
          "sections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SeatMapSectionDto"
            }
          }
        },
        "required": [
          "sections"
        ]
      },
      "SeatMapSectionMetaDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "yStart": {
            "type": "number"
          },
          "rows": {
            "type": "number"
          },
          "cols": {
            "type": "number"
          },
          "xOffset": {
            "type": "number"
          }
        },
        "required": [
          "name",
          "yStart",
          "rows",
          "cols",
          "xOffset"
        ]
      },
      "SeatMapMetaDto": {
        "type": "object",
        "properties": {
          "maxCols": {
            "type": "number"
          },
          "totalRows": {
            "type": "number"
          },
          "sections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SeatMapSectionMetaDto"
            }
          }
        },
        "required": [
          "maxCols",
          "totalRows",
          "sections"
        ]
      },
      "SeatMapTierDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "priceSatang": {
            "type": "number"
          },
          "remaining": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "priceSatang",
          "remaining"
        ]
      },
      "SeatInfoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "section": {
            "type": "string"
          },
          "rowLabel": {
            "type": "string"
          },
          "number": {
            "type": "number"
          },
          "x": {
            "type": "number"
          },
          "y": {
            "type": "number"
          },
          "ticketTypeId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "sold",
              "held",
              "available"
            ]
          },
          "mine": {
            "type": "boolean"
          },
          "expiresAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "section",
          "rowLabel",
          "number",
          "x",
          "y",
          "ticketTypeId",
          "status",
          "mine"
        ]
      },
      "SeatMapDataDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "template": {
            "type": "string"
          },
          "meta": {
            "$ref": "#/components/schemas/SeatMapMetaDto"
          },
          "tiers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SeatMapTierDto"
            }
          },
          "seats": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SeatInfoDto"
            }
          }
        },
        "required": [
          "id",
          "template",
          "meta",
          "tiers",
          "seats"
        ]
      },
      "OrderItemInputDto": {
        "type": "object",
        "properties": {
          "ticketTypeId": {
            "type": "string"
          },
          "quantity": {
            "type": "number",
            "minimum": 1,
            "maximum": 20
          }
        },
        "required": [
          "ticketTypeId",
          "quantity"
        ]
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "items": {
            "minItems": 1,
            "maxItems": 5,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemInputDto"
            }
          },
          "seatIds": {
            "minItems": 1,
            "maxItems": 8,
            "description": "Seat ids currently held by the caller",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "buyerEmail": {
            "type": "string",
            "format": "email",
            "example": "ada@example.com"
          },
          "buyerName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "example": "Ada Lovelace"
          }
        },
        "required": [
          "buyerEmail",
          "buyerName"
        ]
      },
      "EventSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "venueName": {
            "type": "string"
          },
          "startsAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "venueName",
          "startsAt"
        ]
      },
      "TicketTypeRefDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "OrderItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "unitPriceSatang": {
            "type": "number"
          },
          "ticketType": {
            "$ref": "#/components/schemas/TicketTypeRefDto"
          }
        },
        "required": [
          "id",
          "quantity",
          "unitPriceSatang",
          "ticketType"
        ]
      },
      "SeatLabelDto": {
        "type": "object",
        "properties": {
          "section": {
            "type": "string"
          },
          "rowLabel": {
            "type": "string"
          },
          "number": {
            "type": "number"
          }
        },
        "required": [
          "section",
          "rowLabel",
          "number"
        ]
      },
      "OrderTicketDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "qrToken": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "issued",
              "checked_in",
              "void"
            ]
          },
          "attendeeName": {
            "type": "string"
          },
          "ticketType": {
            "$ref": "#/components/schemas/TicketTypeRefDto"
          },
          "seat": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/SeatLabelDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "qrToken",
          "status",
          "attendeeName",
          "ticketType",
          "seat"
        ]
      },
      "OrderPaymentDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "succeeded",
              "failed",
              "requires_action"
            ]
          },
          "checkoutUrl": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "checkoutUrl"
        ]
      },
      "OrderDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "awaiting_payment",
              "paid",
              "expired",
              "canceled",
              "partially_refunded",
              "refunded"
            ],
            "type": "string"
          },
          "totalSatang": {
            "type": "number"
          },
          "refundedSatang": {
            "type": "number"
          },
          "guestToken": {
            "type": "string"
          },
          "buyerEmail": {
            "type": "string"
          },
          "buyerName": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "nullable": true
          },
          "event": {
            "$ref": "#/components/schemas/EventSummaryDto"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemDto"
            }
          },
          "tickets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderTicketDto"
            }
          },
          "payment": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/OrderPaymentDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "status",
          "totalSatang",
          "refundedSatang",
          "guestToken",
          "buyerEmail",
          "buyerName",
          "createdAt",
          "expiresAt",
          "event",
          "items",
          "tickets",
          "payment"
        ]
      },
      "MyTicketDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "qrToken": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "issued",
              "checked_in",
              "void"
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "event": {
            "$ref": "#/components/schemas/EventSummaryDto"
          },
          "ticketType": {
            "$ref": "#/components/schemas/TicketTypeRefDto"
          },
          "seat": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/SeatLabelDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "qrToken",
          "status",
          "createdAt",
          "event",
          "ticketType",
          "seat"
        ]
      },
...
    }
  }
}